+++ client/src/fileutils.c
@@ -1645,7 +1645,7 @@ static int filelist(const char *path, const char *ext, uint8_t last, bool tentat
 
     PrintAndLogEx(NORMAL, "%s── "_GREEN_("%s"), last ? "└" : "├", &path[strip]);
 
-    for (uint16_t i = 0; i < n; i++) {
+    for (int i = 0; i < n; i++) {
 
         char tmp_fullpath[1024] = {0};
         strncat(tmp_fullpath, path, sizeof(tmp_fullpath) - 1);
